Implementing Travis-CI base and new .jar without Debug
This is the start of Travis-CI in this repo because we need to test every PR in case there are any, if the Travis compilation fails, the PR is closed until is updated. Also removed the Debug.java file because we don't want to get the Debug as part of our library.