Updated README
Now shows the real travis-ci status. So if someone sends us a PR, we can check if that PR is actually compiling without errors.

Remember, this project was started with NetBeans, so that's why we give support for NetBeans on the base, but, in the code, we give support for Java in general. So you only need the JAR file and you should not have any problem in any IDE/Compiler.